This also removes some of the optimization using the reciprocals.

```julia
julia> Gray(reinterpret(N0f8, 0x5)) / reinterpret(N0f8, 0x6)
Gray{N0f8}(0.831)     # this PR
ERROR: ArgumentError: # v0.9.4

julia> RGB(reinterpret(N0f8, 0x5)) / reinterpret(N0f8, 0x6)
RGB{N0f8}(0.831,0.831,0.831) # this PR
RGB{N0f8}(0.835,0.835,0.835) # v0.9.4

julia> Gray(eps(N0f8)) / 3.0f0
Gray{Float32}(0.0013071896f0) # this PR
Gray{Float32}(0.0013071897f0) # v0.9.4
```
In the case of `RGB{N0f8}`, the absolute error does not change, but the round-to-even  is applied in this PR.

Fixes #154