@kimikage 2021-05-03
### Benchmarks

<details>
<summary>Script</summary>

```julia
using ColorVectorSpace, ColorTypes, FixedPointNumbers, BenchmarkTools

BenchmarkTools.DEFAULT_PARAMETERS.seconds = 10
N = 1000

g_n0f8  = rand(Gray{N0f8}, N, N);
g_n0f16 = rand(Gray{N0f16}, N, N);
g_f32   = rand(Gray{Float32}, N, N);
g_f64   = rand(Gray{Float64}, N, N);

r_n0f8  = fill(1N0f8, N, N);
r_f64   = fill(1.0, N, N);
r_int   = fill(1, N, N);

rgb_n0f8  = rand(RGB{N0f8}, N, N);
rgb_n0f16 = rand(RGB{N0f16}, N, N);
rgb_f32   = rand(RGB{Float32}, N, N);
rgb_f64   = rand(RGB{Float64}, N, N);

for r in (1N0f8, 1.0, 1)
    @show r
    @btime $g_n0f8  ./ $r;
    @btime $g_n0f16 ./ $r;
    @btime $g_f32   ./ $r;
    @btime $g_f64   ./ $r;
end

for r in (r_n0f8, r_f64, r_int)
    @show typeof(r)
    @btime $g_n0f8  ./ $r;
    @btime $g_n0f16 ./ $r;
    @btime $g_f32   ./ $r;
    @btime $g_f64   ./ $r;
end

for r in (1N0f8, 1.0, 1)
    @show r
    @btime $rgb_n0f8  ./ $r;
    @btime $rgb_n0f16 ./ $r;
    @btime $rgb_f32   ./ $r;
    @btime $rgb_f64   ./ $r;
end

for r in (r_n0f8, r_f64, r_int)
    @show typeof(r)
    @btime $rgb_n0f8  ./ $r;
    @btime $rgb_n0f16 ./ $r;
    @btime $rgb_f32   ./ $r;
    @btime $rgb_f64   ./ $r;
end
```
</details>

```julia
julia> versioninfo()
Julia Version 1.6.1
Commit 6aaedecc44 (2021-04-23 05:59 UTC)
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-11.0.1 (ORCJIT, skylake)
```

| operation          | v0.9.4    | before   | after    |                  |
|:-------------------|:---------:|:--------:|:--------:|:----------------:|
|`g_n0f8   ./ 1N0f8` |( 9.345 ms)|(0.069 ms)| 1.505 ms |                  |
|`g_n0f16  ./ 1N0f8` |( 9.938 ms)|(0.312 ms)| 1.703 ms |                  |
|`g_f32    ./ 1N0f8` |( 0.640 ms)|(0.630 ms)| 0.615 ms |                  |
|`g_f64    ./ 1N0f8` |( 1.437 ms)|(1.472 ms)| 1.464 ms |                  |
|`g_n0f8   ./ 1.0`   |  1.152 ms | 1.136 ms | 1.165 ms |                  |
|`g_n0f16  ./ 1.0`   |  1.196 ms | 1.188 ms | 1.236 ms |                  |
|`g_f32    ./ 1.0`   |  1.257 ms | 1.233 ms | 1.264 ms |                  |
|`g_f64    ./ 1.0`   |  1.418 ms | 1.442 ms | 1.474 ms |                  |
|`g_n0f8  ./ 1`      |  0.592 ms | 0.577 ms | 0.586 ms |                  |
|`g_n0f16 ./ 1`      |  0.598 ms | 0.609 ms | 0.604 ms |                  |
|`g_f32   ./ 1`      |  0.616 ms | 0.662 ms | 0.652 ms |                  |
|`g_f64   ./ 1`      |  1.437 ms | 1.480 ms | 1.431 ms |                  |
|`g_n0f8   ./ r_n0f8`|( 9.859 ms)|(3.105 ms)| 1.689 ms |:heavy_check_mark:|
|`g_n0f16  ./ r_n0f8`|(10.336 ms)|(3.825 ms)| 2.034 ms |:heavy_check_mark:|
|`g_f32    ./ r_n0f8`|( 3.492 ms)|(3.508 ms)| 0.623 ms |:heavy_check_mark:|
|`g_f64    ./ r_n0f8`|( 4.190 ms)|(4.055 ms)| 1.444 ms |:heavy_check_mark:|
|`g_n0f8   ./ r_f64` |  1.520 ms | 1.505 ms | 1.470 ms |                  |
|`g_n0f16  ./ r_f64` |  1.514 ms | 1.533 ms | 1.607 ms |                  |
|`g_f32    ./ r_f64` |  1.632 ms | 1.655 ms | 1.613 ms |                  |
|`g_f64    ./ r_f64` |  1.714 ms | 1.768 ms | 1.794 ms |                  |
|`g_n0f8  ./ r_int`  |  1.384 ms | 1.403 ms | 1.428 ms |                  |
|`g_n0f16 ./ r_int`  |  1.435 ms | 1.381 ms | 1.464 ms |                  |
|`g_f32   ./ r_int`  |  1.324 ms | 1.380 ms | 1.393 ms |                  |
|`g_f64   ./ r_int`  |  1.827 ms | 1.851 ms | 1.878 ms |                  |


| operation            | v0.9.4    | before    | after    |                  |
|:---------------------|:---------:|:---------:|:--------:|:----------------:|
|`rgb_n0f8   ./ 1N0f8` |  9.660 ms |  9.840 ms |14.702 ms |:x:               |
|`rgb_n0f16  ./ 1N0f8` | 10.225 ms |  9.997 ms |14.718 ms |:x:               |
|`rgb_f32    ./ 1N0f8` |(14.172 ms)|(13.965 ms)| 2.678 ms |:heavy_check_mark:|
|`rgb_f64    ./ 1N0f8` |(15.899 ms)|(16.272 ms)| 5.213 ms |:heavy_check_mark:|
|`rgb_n0f8   ./ 1.0`   |  3.711 ms |  3.756 ms | 3.688 ms |                  |
|`rgb_n0f16  ./ 1.0`   |  3.932 ms |  3.933 ms | 3.829 ms |                  |
|`rgb_f32    ./ 1.0`   |  4.144 ms |  4.197 ms | 4.298 ms |                  |
|`rgb_f64    ./ 1.0`   |  4.581 ms |  4.711 ms | 4.747 ms |                  |
|`rgb_n0f8  ./ 1`      |  2.137 ms |  2.078 ms | 2.015 ms |                  |
|`rgb_n0f16 ./ 1`      |  2.697 ms |  2.676 ms | 2.222 ms |                  |
|`rgb_f32   ./ 1`      |  2.533 ms |  2.516 ms | 2.532 ms |                  |
|`rgb_f64   ./ 1`      |  4.611 ms |  4.714 ms | 4.641 ms |                  |
|`rgb_n0f8   ./ r_n0f8`|  9.954 ms | 10.004 ms |14.264 ms |:x:               |
|`rgb_n0f16  ./ r_n0f8`| 10.561 ms | 10.454 ms |14.975 ms |:x:               |
|`rgb_f32    ./ r_n0f8`|(14.058 ms)|(13.961 ms)| 2.709 ms |:heavy_check_mark:|
|`rgb_f64    ./ r_n0f8`|(15.494 ms)|(16.034 ms)| 4.869 ms |:heavy_check_mark:|
|`rgb_n0f8   ./ r_f64` |  4.067 ms |  4.006 ms | 4.031 ms |                  |
|`rgb_n0f16  ./ r_f64` |  4.262 ms |  3.958 ms | 4.172 ms |                  |
|`rgb_f32    ./ r_f64` |  4.415 ms |  4.383 ms | 4.330 ms |                  |
|`rgb_f64    ./ r_f64` |  4.992 ms |  4.865 ms | 5.054 ms |                  |
|`rgb_n0f8  ./ r_int`  |  2.850 ms |  2.590 ms | 2.594 ms |                  |
|`rgb_n0f16 ./ r_int`  |  3.458 ms |  3.347 ms | 3.103 ms |                  |
|`rgb_f32   ./ r_int`  |  2.873 ms |  2.568 ms | 2.659 ms |                  |
|`rgb_f64   ./ r_int`  |  5.059 ms |  4.649 ms | 4.585 ms |                  |