@kimikage 2021-05-03
As shown above, this will result in slowdowns in some cases, but if the speed performance is really critical, I don't think you should use the "checked" division.