@kimikage 2021-05-04
If we find no issues other than slowdowns in the edge cases, I will merge this in this weekend. Then, I'll update PR #153 and release v0.9.5 in the next week.