@kimikage 2021-05-08
Since this is intended to fix a bug, I'll merge this.
If you have problems with slowdowns, please report them. In general, lookup tables are effective for `N0f8` calculations.